requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==1.10.17
email-validator==2.2.0
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==10.4.0
python-multipart==0.0.9
pymongo[srv]==4.8.0
openpyxl==3.1.5